requirements.txt
blinker==1.7.0
click==8.1.7
cowsay==6.1
Flask==3.0.1
fortune-python==1.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
Werkzeug==3.0.1
